Divide whiteboard interactions in two
