+++ spec/suitec/impact_studio_visualizations_spec.rb
@@ -40,7 +40,7 @@ describe 'The Impact Studio', order: :defined do
     @asset_library = Page::SuiteCPages::AssetLibraryDetailPage.new @driver
     @engagement_index = Page::SuiteCPages::EngagementIndexPage.new @driver
     @impact_studio = Page::SuiteCPages::ImpactStudioPage.new @driver
-    @whiteboards = Page::SuiteCPages::WhiteboardsPage.new @driver
+    @whiteboards = Page::SuiteCPages::WhiteboardPage.new @driver
 
     # Create course site if necessary
     @canvas.log_in(@cal_net, Utils.super_admin_username, Utils.super_admin_password)
