+++ spec/suitec/whiteboard_assets_spec.rb
@@ -22,7 +22,7 @@ describe 'Whiteboard Add Asset', order: :defined do
     @cal_net = Page::CalNetPage.new @driver
     @asset_library = Page::SuiteCPages::AssetLibraryDetailPage.new @driver
     @engagement_index = Page::SuiteCPages::EngagementIndexConfigPage.new @driver
-    @whiteboards = Page::SuiteCPages::WhiteboardsPage.new @driver
+    @whiteboards = Page::SuiteCPages::WhiteboardPage.new @driver
 
     # Create test course
     @canvas.log_in(@cal_net, (event.actor = @admin).username, Utils.super_admin_password)
